React steps
