+++ identity/app/views/completeConsents.scala.html
@@ -76,7 +76,7 @@
         }
         <aside class="identity-forms-message__body">
             <a class="manage-account__button manage-account__button--secondary" data-link-name="complete-consents : cta-bottom" href="@LinkTo(returnUrl)">
-                Go to the theguardian.com
+                Go to The Guardian
                 @fragments.inlineSvg("arrow-right", "icon")
             </a>
         </aside>
